+++ processor/src/test/resources/expected/com/example/repository/InjectEnabledRepositoryRepository.java
@@ -10,7 +10,6 @@ import javax.inject.Singleton;
 
 @Generated("arez.processor.ArezProcessor")
 @ArezComponent(
-    nameIncludesId = false,
     inject = Feature.ENABLE
 )
 @Singleton