+++ processor/src/test/resources/expected/com/example/repository/RepositoryWithExplicitIdRepository.java
@@ -9,9 +9,7 @@ import javax.annotation.Nullable;
 import javax.inject.Singleton;
 
 @Generated("arez.processor.ArezProcessor")
-@ArezComponent(
-    nameIncludesId = false
-)
+@ArezComponent
 @Singleton
 public abstract class RepositoryWithExplicitIdRepository extends AbstractRepository<Integer, RepositoryWithExplicitId, RepositoryWithExplicitIdRepository> {
   RepositoryWithExplicitIdRepository() {